fixed the all the syntax errors
This should be able to be ran - however still not 100% on everything especially the way I am sending data out the buffer and managing the ring bufffer. I have it such that the head is the oldest data and the tail is the newest data. When sending it out it will start at the tail and decrement until it reaches the head